sharq/sha: fix sign extension

+++ Jerry.cpp
@@ -1190,7 +1190,7 @@ void Jerry::compute()
   case DSPI::SHA:
     if ( mStageWrite.canUpdateReg() )
     {
-      mStageWrite.updateReg( mStageCompute.regDst, mStageCompute.dataSrc > 0 ? ( uint32_t )( ( int64_t )( mStageCompute.dataDst ) >> mStageCompute.dataSrc ) : mStageCompute.dataDst << mStageCompute.dataSrc );
+      mStageWrite.updateReg( mStageCompute.regDst, mStageCompute.dataSrc > 0 ? ( uint32_t )( ( int64_t )(int32_t)( mStageCompute.dataDst ) >> mStageCompute.dataSrc ) : mStageCompute.dataDst << mStageCompute.dataSrc );
       mStageWrite.regFlags.z = mStageWrite.data == 0 ? 1 : 0;
       mStageWrite.regFlags.n = mStageWrite.data >> 31;
       mStageWrite.regFlags.c = mStageCompute.dataSrc > 0 ? ( mStageCompute.dataDst & 1 ) : ( mStageCompute.dataDst >> 31 );
@@ -1202,7 +1202,7 @@ void Jerry::compute()
   case DSPI::SHARQ:
     if ( mStageWrite.canUpdateReg() )
     {
-      mStageWrite.updateReg( mStageCompute.regDst, ( uint32_t )( ( int64_t )( mStageCompute.dataDst ) >> mStageCompute.dataSrc ) );
+      mStageWrite.updateReg( mStageCompute.regDst, ( uint32_t )( ((int64_t)(int32_t)mStageCompute.dataDst ) >> mStageCompute.dataSrc ) );
       mStageWrite.regFlags.z = mStageWrite.data == 0 ? 1 : 0;
       mStageWrite.regFlags.n = mStageWrite.data >> 31;
       mStageWrite.regFlags.c = mStageCompute.dataDst & 1;
